BabySitter: Do not get the child listing of a deleted node
If the node is deleted and the watcher is triggered it would try to list the
child nodes but a NoNodeException would be thrown causing the voltdb instance
to crash. When NodeDeleted event is received do not list the children.

Also, the comparison in the watch has a comment that says string comparison
doesn't work and then proceeds to do a string comparison. Update the
comparison to actually be an int comparison.